+++ graphs.py
@@ -65,7 +65,6 @@ def bar_plot(data, options):
         "margin_t": options.pop("margin_t"),
         "margin_b": options.pop("margin_b"),
     }
-    labels = options.pop("labels")
     subplot_label = options.pop("subplot_label")
 
     fig_options = ChainMap(
@@ -88,10 +87,7 @@ def bar_plot(data, options):
             flash(f"Scalar plot error: {ve}", category="error")
         raise PlottingError(f"Scalar plot error: {ve}")
 
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
+    # Plot Labels:
     if subplot_label:
         fig.for_each_annotation(lambda a: a.update(text=f"{subplot_label} {a.text.split('=')[-1]}"))
     else:
@@ -140,7 +136,6 @@ def bar_plot(data, options):
 
 def radar_plot(data, options):
     axis_title = options.pop("axis_title") or add_unit_to_label(options["r"], data)
-    labels = options.pop("labels")
 
     categories = data[options["theta"]].unique()
 
@@ -154,11 +149,6 @@ def radar_plot(data, options):
                 name=ellipse
             )
         )
-
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
 
     fig.update_layout(
         polar={
@@ -181,7 +171,6 @@ def dot_plot(data, options):
     y = data[options["y"]].unique()
     xaxis_title = options.pop("xaxis_title") or add_unit_to_label(options["x"], data)
     legend_title = options.pop("legend_title")
-    labels = options.pop("labels")
 
     fig = go.Figure()
 
@@ -192,11 +181,6 @@ def dot_plot(data, options):
             y=y,
             name=category,
         ))
-
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
 
     fig.update_traces(mode='markers', marker=dict(line_width=1, symbol='circle', size=16))
     fig.update_layout(
@@ -223,7 +207,7 @@ def line_plot(data, options):
     xaxis_title = options.pop("xaxis_title") or "Timeindex"
     yaxis_title = options.pop("yaxis_title") or add_unit_to_label("", data)
     legend_title = options.pop("legend_title")
-    labels = options.pop("labels")
+
     fig_options = ChainMap(
         options,
         GRAPHS_DEFAULT_OPTIONS["timeseries"]["line"].get_defaults(exclude_non_plotly_options=True)
@@ -253,11 +237,6 @@ def line_plot(data, options):
         }
     )
 
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
-
     fig.update_layout(
         yaxis_title=yaxis_title,
         xaxis_title=xaxis_title,
@@ -274,7 +253,6 @@ def box_plot(data, options):
     xaxis_title = options.pop("xaxis_title") or "Time"
     yaxis_title = options.pop("yaxis_title")
     legend_title = options.pop("legend_title")
-    labels = options.pop("labels")
     sample = options.pop("sample")
     fig_options = ChainMap(
         options,
@@ -299,11 +277,6 @@ def box_plot(data, options):
         flash(f"Timeseries plot error: {ve}", category="error")
         raise PlottingError(f"Timeseries plot error: {ve}")
 
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
-
     fig.update_layout(
         xaxis_title=xaxis_title,
         yaxis_title=yaxis_title or add_unit_to_label(fig_options["y"], ts_flattened),
@@ -322,7 +295,6 @@ def heat_map(data, options):
     xaxis_title = options.pop("xaxis_title") or x
     yaxis_title = options.pop("yaxis_title") or y
     legend_title = options.pop("legend_title") or add_unit_to_label("Value", data)
-    labels = options.pop("labels")
 
     fig_options = ChainMap(
         options,
@@ -352,11 +324,6 @@ def heat_map(data, options):
     except ValueError as ve:
         flash(f"Timeseries plot error: {ve}", category="error")
         raise PlottingError(f"Timeseries plot error: {ve}")
-
-    # Override labels:
-    if labels:
-        for trace in fig.data:
-            trace.name = labels.get(trace.name, trace.name)
 
     fig.update_xaxes(side="top")
     fig.update_layout(