+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.12.0] - 2021-08-30
 ### Added
 - option to hide legend
 - option for logarithmic y-axis
@@ -13,7 +13,7 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - logarithmic axes in all subplots
 - logarithmic axis starts at zero
 
-## [0.11.0]
+## [0.11.0] - 2021-08-27
 ### Added
 - custom labels for subplots